dsymbol.modulecache: Ignore errors when querying file attributes
std.file.exists is insufficient to ensure that querying the attributes
at that path will succeed.  Not only is doing so a TOCTOU bug, but,
std.file.exists performs an lstat call, whereas isFile/isDir performs
a stat call (which we do need to do to support symlinks); this will
result in an exception when the path is a broken symlink, as the
former will succeed while the latter will fail.

Fix this by removing explicit std.file.exists checks, and replacing
them with a wrapper which catches and handles error conditions,
treating them as the path not satisfying the desired property.